+++ src/cryptonote/tx/cryptonote_tx_utils.cpp
@@ -457,7 +457,7 @@ namespace cryptonote
   bool generate_genesis_block(
       block& bl
     , std::string_view const & genesis_tx
-    , uint32_t nonce
+    , uint64_t nonce
     )
   {
     //genesis block
