+++ src/cryptonote/tx/cryptonote_tx_utils.h
@@ -121,7 +121,7 @@ namespace cryptonote
   bool generate_genesis_block(
       block& bl
     , std::string_view const & genesis_tx
-    , uint32_t nonce
+    , uint64_t nonce
     );
 
   class Blockchain;
